Support jinja2 3.1
Since Jinja2 3.1, `while_` and `autoescape` have been removed. They've
been included in stdlib for a while and didn't need to be included.

Unfortuantely we don't pin to any lower or upper bounds of jinja2, so
the best/fastest option is to just version detect and include by default
if less than 3.1.

Fixes #100
Fixes #101